block: clarify the buffer-handling constraints for read, write
Each of the page_aligned_buffers must be a whole number of sectors in
size. There's no limit to the number of sectors per buffer or the number
of buffers per request.

I considered adding a notion of 'async' vs 'sync', but we already allow
issuing async operations (instances of 'a io), and synchronising with
them ('bind'). The plan is to start with O_DIRECT-like I/O at the bottom
and then consider adding buffering later. It might be better to always
perform buffering at the application-level, rather than having a
one-size-fits-all policy in the block 'layer'.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>